Fix double-free in test_http_transport and C++ casts in security tests
test_http_transport.cc:
- Remove redundant axiom_node_free_tree() call after axis2_json_reader_free()
- axis2_json_reader_free() now frees the axiom tree internally (memory leak fix)
- Fixes use-after-free detected by AddressSanitizer

test_buffer_safety.cc / test_json_security.cc:
- Add explicit C++ casts for void* returns from axutil functions
- Fix signedness comparison warning

+++ test/core/security/test_json_security.cc
@@ -209,7 +209,7 @@ TEST_F(JSONSecurityTest, ArrayHandling) {
     json_object *obj = json_tokener_parse(array_json);
     ASSERT_NE(obj, nullptr) << "JSON array should parse";
     EXPECT_TRUE(json_object_is_type(obj, json_type_array));
-    EXPECT_EQ(json_object_array_length(obj), 5);
+    EXPECT_EQ(json_object_array_length(obj), (size_t)5);
 
     json_object_put(obj);
 }
